requirements.txt
appnope==0.1.0
coloredlogs==5.0
decorator==4.0.10
humanfriendly==1.44.7
ipdb==0.10.1
ipython==5.1.0
ipython-genutils==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.7
ptyprocess==0.5.1
Pygments==2.1.3
pyserial==3.1.1
requests==2.11.1
simplegeneric==0.8.1
six==1.10.0
traitlets==4.2.2
wcwidth==0.1.7
